requirements.txt
google-cloud-logging==3.2.5
twilio==7.15.3
requests==2.25.1
